ebpms-break-health-calculation-master/requirements.txt
Click==7.0
cx-Oracle==7.3.0
Flask==1.1.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
kazoo==2.5.0
MarkupSafe==1.1.1
prometheus-flask-exporter==0.12.1
psutil==5.8.0
pytz==2019.3
six==1.13.0
tabulate==0.8.6
Werkzeug==0.16.0
statsmodels==0.12.2
numpy==1.17.2
pandas==1.1.2
psycopg2-binary
